requirements.txt
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Deprecated==1.2.14
docopt==0.6.2
idna==3.4
pycparser==2.21
PyGithub==1.59.1
PyJWT==2.8.0
PyNaCl==1.5.0
PyYAML==6.0.1
requests==2.31.0
tabulate==0.9.0
urllib3==2.0.4
wrapt==1.15.0
